requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
boto3==1.26.134
botocore==1.29.134
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
jmespath==1.0.1
multidict==6.0.4
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.30.0
s3transfer==0.6.1
six==1.16.0
urllib3==2.0.2
yarl==1.9.2
